Initial commit
Made in about 17 hours over night with some procrastination in the form of watching streams and chatting with people. The blessed library is a bit crap when it comes to usability and the TypeScript types, so that bit of the code is pretty meh, but it works. The pactl wrapper is pretty cool though. Quite proud of the list parser!

Usage is simple: run the cli, select sink input on the left, press tab, select a sink to move it to on the right. Press c or n to create new modules to allow routing the audio in funky ways. Press d to easily delete those created modules. It's simple but it works (ish) and that's all I wanted.